+++ src/consumer/consumer.js
@@ -344,7 +344,7 @@ export default class Consumer {
     silentPackageManagerResult,
     componentsWithDependencies,
     writeToPath,
-    force = true,
+    force = true, // override files
     writePackageJson = true,
     writeBitJson = true,
     writeBitDependencies = false,